+++ src/mainwindow.h
@@ -67,7 +67,7 @@ public:
             QString name, QString text,
             CreateNewNoteOptions options = CreateNewNoteOption::None);
 
-    void openSettingsDialog(int page = 0);
+    void openSettingsDialog(int page = 0, bool openScriptRepository = false);
 
     void restoreTrashedNoteOnServer(QString fileName, int timestamp);
 
@@ -458,6 +458,10 @@ private slots:
     void on_tagTreeWidget_itemCollapsed(QTreeWidgetItem *item);
 
     void on_tagTreeWidget_itemExpanded(QTreeWidgetItem *item);
+
+    void on_actionScript_repository_triggered();
+
+    void on_actionScript_settings_triggered();
 
 private:
     Ui::MainWindow *ui;